% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-data.R
\docType{data}
\name{hr_test}
\alias{hr_test}
\title{Decision tree test data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 441 rows and 35 columns.}
\usage{
hr_test
}
\description{
Test data set for decision tree model
}
\keyword{datasets}
