% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-data.R
\docType{data}
\name{hr_train}
\alias{hr_train}
\title{Decision tree train data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1029 rows and 35 columns.}
\usage{
hr_train
}
\description{
Training data set for decision tree model
}
\keyword{datasets}
