% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbar-prep-data.R
\name{mbar_prep_data}
\alias{mbar_prep_data}
\title{Prepare Data}
\usage{
mbar_prep_data(data, id, products)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{id}{Unique id for the transaction.}

\item{products}{items part of the transaction.}
}
\description{
Prepares transaction data for analysis.
}
\examples{
mbar_prep_data(mba_sample, InvoiceNo, Description)

}
