% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal-cp.R
\name{optimal_cp}
\alias{optimal_cp}
\title{Optimal complexity parameter}
\usage{
optimal_cp(model)
}
\arguments{
\item{model}{An object of class \code{rpart}.}
}
\description{
Extract optimal complexity parameter for tree pruning.
}
\examples{
model <- rpart::rpart(Species ~ ., data = iris)
optimal_cp(model)

}
