% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{plot_roc}
\alias{plot_roc}
\alias{plot_prec_rec}
\alias{plot_sens_spec}
\alias{plot_lift_curve}
\title{Plot validation curves}
\usage{
plot_roc(perform, line_color = "blue")

plot_prec_rec(perform, line_color = "blue")

plot_sens_spec(perform, line_color = "blue")

plot_lift_curve(perform, line_color = "blue")
}
\arguments{
\item{perform}{An object of class \code{prediction}.}

\item{line_color}{Color of lines in the plot.}
}
\description{
Plot curves to validate decision tree model.
}
\examples{
model   <- rpart::rpart(Attrition ~ ., data = hr_train)
perform <- tree_prediction(model, hr_test, Attrition)
plot_roc(perform)
plot_prec_rec(perform)
plot_sens_spec(perform)
plot_lift_curve(perform)

}
