% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{tree_auc}
\alias{tree_auc}
\title{AUC}
\usage{
tree_auc(perform)
}
\arguments{
\item{perform}{An object of class \code{prediction}.}
}
\description{
Area under the curve.
}
\examples{
model   <- rpart::rpart(Attrition ~ ., data = hr_train)
perform <- tree_prediction(model, hr_test, Attrition)
tree_auc(perform)

}
