% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{tree_prediction}
\alias{tree_prediction}
\title{Class probability prediction}
\usage{
tree_prediction(model, test_data, response)
}
\arguments{
\item{An}{object of class \code{rpart}.}

\item{A}{\code{data.frame} or \code{tibble}.}

\item{Response}{variable.}
}
\description{
Predict class probability on a data set.
}
\examples{
model <- rpart::rpart(Attrition ~ ., data = hr_train)
tree_prediction(model, hr_test, Attrition)

}
